% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unique_ids.R
\name{create_unique_ids}
\alias{create_unique_ids}
\title{create_unique_ids}
\usage{
create_unique_ids(n = 1, char_len = 5)
}
\arguments{
\item{n}{integer}

\item{char_len}{integer}
}
\value{
character string
}
\description{
create unique ids with a combo of alpha and numeric characters
}
